+++ src/test/java/woowa/camp/jspcafe/repository/UserRepositoryTest.java
@@ -17,7 +17,7 @@ import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
 import woowa.camp.jspcafe.repository.user.DBUserRepository;
 import woowa.camp.jspcafe.repository.user.UserRepository;
 import woowa.camp.jspcafe.utils.FixedDateTimeProvider;
-import woowa.camp.jspcafe.utils.time.DateTimeProvider;
+import woowa.camp.jspcafe.infra.time.DateTimeProvider;
 
 @ExtendWith(UserDBSetupExtension.class)
 class UserRepositoryTest {
